+++ R/get_priors.R
@@ -95,6 +95,10 @@ get_priors.brmsfit <- function(x, ...) {
 
 
   prior_info <- x$prior[x$prior$coef != "" & x$prior$class %in% c("b", "(Intercept)"), ]
+  # find additional components, avoid duplicated coef-names
+  components <- prior_info$dpar != ""
+  prior_info$dpar[components] <- paste0(prior_info$dpar[components], "_")
+  prior_info$coef <- paste0(prior_info$dpar, prior_info$coef)
 
   prior_info$Distribution <- gsub("(.*)\\(.*", "\\1", prior_info$prior)
   prior_info$Location <- gsub("(.*)\\((.*)\\,(.*)", "\\2", prior_info$prior)