+++ R/is_model_supported.R
@@ -51,6 +51,7 @@ supported_models <- function() {
     "bigglm",
     "blavaan",
     "bracl",
+    "brglm",
     "brmsfit",
     "brmultinom",
     "censReg",